rename common package to constant
